// model/chatModel.ts
import mongoose, { Schema } from "mongoose";

export interface IChat {
  _id?: string; // Optional if generated by MongoDB
  senderEmail: string;
  message: string;
  timestamp?: Date; // Optional if set by backend
}

const chatSchema = new mongoose.Schema({
  senderEmail: { type: String, required: true },
  message: { type: String, required: true },
  timestamp: { type: Date, default: Date.now },
});

const Chat = mongoose.model<IChat>("Chat", chatSchema);

export default Chat;